requirements.txt
Flask==3.1.2
gunicorn
groq==0.34.0
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-groq==0.1.3
pypdf==6.2.0
python-dotenv==1.2.1
sentence-transformers
